+++ test/session_test/cookies_test.rb
@@ -64,6 +64,21 @@ module SessionTest
         assert_equal false, session.httponly
       end
 
+      def test_same_site
+        assert_nil UserSession.same_site
+        assert_nil UserSession.new.same_site
+
+        UserSession.same_site 'Strict'
+        assert_equal 'Strict', UserSession.same_site
+        session = UserSession.new
+        assert_equal 'Strict', session.same_site
+        session.same_site = 'Lax'
+        assert_equal 'Lax', session.same_site
+
+        assert_raise(ArgumentError) { UserSession.same_site 'foo' }
+        assert_raise(ArgumentError) { UserSession.new.same_site 'foo' }
+      end
+
       def test_sign_cookie
         UserSession.sign_cookie = true
         assert_equal true, UserSession.sign_cookie
@@ -184,6 +199,16 @@ module SessionTest
             controller.cookies["user_credentials"]
           )
         end
+      end
+
+      def test_after_save_save_cookie_with_same_site
+        session = UserSession.new(users(:ben))
+        session.same_site = 'Strict'
+        assert session.save
+        assert_equal(
+          'Strict',
+          controller.cookies.set_cookies['user_credentials'][:same_site]
+        )
       end
 
       def test_after_destroy_destroy_cookie